+++ src/main/java/me/grish/veinforge/macro/impl/CommissionMacro/states/StartingState.java
@@ -3,6 +3,7 @@ package me.grish.veinforge.macro.impl.CommissionMacro.states;
 import me.grish.veinforge.VeinForge;
 import me.grish.veinforge.macro.impl.CommissionMacro.CommissionMacro;
 import me.grish.veinforge.util.InventoryUtil;
+import me.grish.veinforge.util.ToolSelector;
 
 import java.util.Objects;
 
@@ -15,11 +16,11 @@ public class StartingState implements CommissionMacroState {
 
     @Override
     public CommissionMacroState onTick(CommissionMacro macro) {
-        if (Objects.equals(VeinForge.config().general.miningTool, "")) {
+        if (Objects.equals(VeinForge.config().general.miningTool, "") && !ToolSelector.isAutoSelectEnabled()) {
             macro.disable("Please set a Mining Tool in the config");
             return null;
         }
-        if (Objects.equals(VeinForge.config().commission.dwarvenCommission.slayerWeapon, "")) {
+        if (Objects.equals(VeinForge.config().commission.dwarvenCommission.slayerWeapon, "") && !ToolSelector.isAutoSelectEnabled()) {
             macro.disable("Please set a Slayer Weapon in the config");
             return null;
         }